requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typer>=0.9.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.10.0
statsmodels>=0.14.0


